requirements_replit.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.6.0
anthropic==0.45.2
anyio==3.7.1
attrs==25.1.0
backend==0.2.4.1
boto3==1.36.21
botocore==1.36.21
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cryptography==44.0.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.112.1
frozenlist==1.5.0
gotrue==2.11.3
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.4
jiter==0.8.2
jmespath==1.0.1
multidict==6.1.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
postgrest==0.19.3
propcache==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.1
realtime==2.3.0
requests==2.32.3
rsa==4.9
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.0
starlette==0.38.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.23.2
websockets==14.2
XlsxWriter==3.2.2
yarl==1.18.3